# backend/docs/swagger.json
{
    "schemes": [
        "http,",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Sila API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workout/add-set/{id}": {
            "post": {
                "description": "This endpoint allows adding a set (reps and weight) to an ongoing workout session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Add a set to a workout session",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Set data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input or session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Workout session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to add set",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workout/complete/{id}": {
            "put": {
                "description": "Marks a workout session as completed and allows adding a final note.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Complete a workout session",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Completion note",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompleteWorkoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout session completed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input or session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Workout session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to complete workout session",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workout/exercise/{exercise_id}": {
            "get": {
                "description": "Retrieve paginated workout sessions for a specific exercise for the authenticated user, with support for different languages.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Get workout sessions by exercise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exercise ID",
                        "name": "exercise_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of workout sessions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid exercise ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve workout sessions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workout/incomplete": {
            "get": {
                "description": "Retrieve all incomplete workout sessions for the authenticated user, with support for different languages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Get incomplete workout sessions",
                "responses": {
                    "200": {
                        "description": "List of incomplete workouts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IncompleteWorkoutResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve incomplete workouts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/workout/start": {
            "post": {
                "description": "This endpoint starts a workout session for a specific exercise and optionally returns the last completed session if available.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Start a workout session",
                "parameters": [
                    {
                        "description": "Exercise ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StartWorkoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New workout session and last session (if any)",
                        "schema": {
                            "$ref": "#/definitions/models.StartWorkoutResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or Exercise ID missing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to start workout session",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workout/{id}": {
            "get": {
                "description": "Retrieve details of a specific workout session including sets, exercise info, and muscle group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workout"
                ],
                "summary": "Get details of a workout session",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout session details",
                        "schema": {
                            "$ref": "#/definitions/models.WorkoutSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Workout session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddSetRequest": {
            "type": "object",
            "properties": {
                "reps": {
                    "type": "integer",
                    "example": 10
                },
                "weight": {
                    "type": "number",
                    "example": 50.5
                }
            }
        },
        "models.CompleteWorkoutRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "type": "string",
                    "example": "Great workout!"
                }
            }
        },
...
        "models.Exercise": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description_eng": {
                    "type": "string"
                },
                "description_ru": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "muscle_group": {
                    "$ref": "#/definitions/models.MuscleGroup"
                },
                "name_eng": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
...
        "models.IncompleteWorkoutResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "exercise": {
                    "$ref": "#/definitions/models.Exercise"
                },
                "id": {
                    "type": "integer"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "note": {
                    "type": "string"
                },
                "sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Set"
                    }
                }
            }
        },
...
        "models.MuscleGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name_eng": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
...
        "models.Set": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reps": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                },
                "workout_session_id": {
                    "type": "integer"
                }
            }
        },
        "models.StartWorkoutRequest": {
            "type": "object",
            "properties": {
                "exercise_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.StartWorkoutResponse": {
            "type": "object",
            "required": [
                "created_at",
                "is_completed",
                "session_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "last_session": {
                    "$ref": "#/definitions/models.WorkoutSessionResponse"
                },
                "session_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.WorkoutSessionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "exercise": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "muscle_group": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "note": {
                    "type": "string"
                },
                "sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Set"
                    }
                }
            }
        }
    }
}